front/requirements.txt
Flask
requests
python-dotenv
psutil